refactor: generate module in style of match module
